+++ src/apis/member/info.py
@@ -68,12 +68,12 @@ async def get_list(request: Request, response: Response, authorization: str = He
     name = convertQuotation(name).lower()
 
     order_by_last_seen = False
-    if order_by not in ['user_id', 'name', 'uid', 'discordid', 'highest_role', 'join_timestamp', 'last_seen']:
-        order_by = "user_id"
+    if order_by not in ['uid', 'userid', 'name', 'discordid', 'steamid', 'truckersmpid', 'highest_role', 'join_timestamp', 'last_seen']:
+        order_by = "userid"
         order = "asc"
     if order_by == "last_seen":
         order_by_last_seen = True
-    cvt = {"user_id": "userid", "name": "name", "uid": "uid", "discordid": "discordid", "join_timestamp": "join_timestamp", "highest_role": "highest_role", "last_seen": "userid"}
+    cvt = {"userid": "userid", "name": "name", "uid": "uid", "discordid": "discordid", "steamid": "steamid", "truckersmpid": "truckersmpid", "join_timestamp": "join_timestamp", "highest_role": "highest_role", "last_seen": "userid"}
     order_by = cvt[order_by]
 
     sort_by_highest_role = False