+++ src/plugins/downloads.py
@@ -57,7 +57,7 @@ async def get_list(request: Request, response: Response, authorization: str = He
     elif page_size >= 250:
         page_size = 250
 
-    if order_by not in ["orderid", "downloadsid", "timestamp", "title", "click_count"]:
+    if order_by not in ["orderid", "downloadsid", "title", "timestamp", "click_count"]:
         order_by = "orderid"
         order = "asc"
     order = order.lower()
@@ -66,7 +66,7 @@ async def get_list(request: Request, response: Response, authorization: str = He
 
     base_rows = 0
     tot = 0
-    await app.db.execute(dhrid, f"SELECT downloadsid FROM downloads WHERE downloadsid >= 0 {limit} ORDER BY is_pinned DESC, {order_by} {order}, timestamp DESC")
+    await app.db.execute(dhrid, f"SELECT downloadsid FROM downloads WHERE downloadsid >= 0 {limit} ORDER BY is_pinned DESC, {order_by} {order}, downloadsid DESC")
     t = await app.db.fetchall(dhrid)
     if len(t) == 0:
         return {"list": [], "total_items": 0, "total_pages": 0}
@@ -78,7 +78,7 @@ async def get_list(request: Request, response: Response, authorization: str = He
             base_rows += 1
         tot -= base_rows
 
-    await app.db.execute(dhrid, f"SELECT downloadsid, userid, title, description, link, click_count, orderid, is_pinned, timestamp FROM downloads WHERE downloadsid >= 0 {limit} ORDER BY is_pinned DESC, {order_by} {order}, timestamp DESC LIMIT {base_rows + max(page-1, 0) * page_size}, {page_size}")
+    await app.db.execute(dhrid, f"SELECT downloadsid, userid, title, description, link, click_count, orderid, is_pinned, timestamp FROM downloads WHERE downloadsid >= 0 {limit} ORDER BY is_pinned DESC, {order_by} {order}, downloadsid DESC LIMIT {base_rows + max(page-1, 0) * page_size}, {page_size}")
     t = await app.db.fetchall(dhrid)
     ret = []
     for tt in t: